tweak some display stuff
darcs-hash:20060619014421-55968-38d4cb3355d08bd62d86fd06befe7c1c9e42ce1e.gz

+++ src/jobs/netcentral.hpp
@@ -26,7 +26,7 @@ class NetCentral : public JobQueue {
 
         fd_set master;   // master file descriptor list
         fd_set read_fds; // temp file descriptor list for select()
-        Uint32 num_read_bytes;
+        double num_read_bytes;
   
         vector<string> job_filenames; 
     private:
